#2473 Refactor MDTImportStringService
